[style] 반응형 추가
