+++ src/component/Navbar.js
@@ -12,11 +12,11 @@ const Navbar = () => {
       <Menus>
         <div className="menu">
           <p>3</p>
-          <div>Tasks</div>
+          <div className="active">Tasks</div>
         </div>
 
         <div className="menu">
-          <p className="yellow">3</p>
+          <p className="yellow">1</p>
           <div>Ideas</div>
         </div>
 
@@ -30,7 +30,7 @@ const Navbar = () => {
           <img src={play} />
         </PlayBtn>
 
-        <p>Play - music</p>
+        <p>Play - Levitating ... </p>
       </PlayBox>
 
       <SearchBox>
@@ -43,56 +43,9 @@ const Navbar = () => {
 
 export default Navbar;
 
-const SearchBox = styled.div`
-  width: 176px;
-  height: 60px;
-
-  background: rgba(217, 217, 217, 0.2);
-  border-radius: 15px;
-
-  display: flex;
-  align-items: center;
-
-  margin-left: 100px;
-  img {
-    width: 14px;
-    height: 14px;
-    margin-left: 28px;
-  }
-
-  p {
-    margin-left: 26px;
-
-    font-family: "Cafe24Ssurround";
-    font-style: normal;
-    font-weight: 700;
-    font-size: 15px;
-    color: #646464;
-  }
-`;
-const PlayBox = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  margin-left: auto;
-
-  p {
-    font-family: "Cafe24Ssurround";
-    font-style: normal;
-    font-weight: 700;
-    font-size: 16px;
-    line-height: 19px;
-
-    color: rgba(0, 0, 0, 0.3);
-
-    margin: 0 0 0 36px;
-  }
-`;
-
 const Logo = styled.img`
-  width: 56px;
-  height: 56px;
+  width: 42px;
+  height: 42px;
 `;
 
 const Navdiv = styled.div`
@@ -104,12 +57,12 @@ const Navdiv = styled.div`
   font-size: 24px;
   line-height: 28px;
 
-  margin: 30px 50px;
+  margin: 20px 50px;
 `;
 
 const Menus = styled.div`
   display: flex;
-  margin-left: 130px;
+  margin-left: 10%;
   width: auto;
 
   .menu {
@@ -117,15 +70,19 @@ const Menus = styled.div`
     display: flex;
     justify-content: center;
     align-items: center;
-    margin-right: 50px;
+    margin-right: 20%;
 
     div {
       font-family: "Cafe24Ssurround";
       font-style: normal;
       font-weight: 700;
-      font-size: 24px;
+      font-size: 20px;
       line-height: 28px;
       color: gray;
+    }
+
+    .active {
+      color: black;
     }
   }
 
@@ -134,8 +91,8 @@ const Menus = styled.div`
     top: 0;
     right: -13px;
 
-    width: 18px;
-    height: 18px;
+    width: 16px;
+    height: 16px;
 
     background: #476ef7;
     border-radius: 50%;
@@ -144,7 +101,7 @@ const Menus = styled.div`
     font-family: "Cafe24Ssurround";
     font-style: normal;
     font-weight: 700;
-    font-size: 10px;
+    font-size: 8px;
 
     margin: 0;
 
@@ -161,9 +118,32 @@ const Menus = styled.div`
   }
 `;
 
+const PlayBox = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin-left: auto;
+
+  p {
+    font-family: "Cafe24Ssurround";
+    font-style: normal;
+    font-weight: 700;
+    font-size: 14px;
+    line-height: 19px;
+
+    color: rgba(0, 0, 0, 0.3);
+
+    margin: 0 0 0 36px;
+  }
+
+  @media screen and (max-width: 1150px) {
+    display: none;
+  }
+`;
+
 const PlayBtn = styled.div`
-  width: 40px;
-  height: 40px;
+  width: 30px;
+  height: 30px;
   border-radius: 50%;
   background-color: #476ef7;
 
@@ -172,8 +152,45 @@ const PlayBtn = styled.div`
   align-items: center;
 
   img {
-    width: 13px;
-    height: 13px;
-    margin-left: 5px;
+    width: 10px;
+    height: 10px;
+    margin-left: 3px;
+  }
+`;
+
+const SearchBox = styled.div`
+  width: 176px;
+  height: 45px;
+
+  background: rgba(217, 217, 217, 0.2);
+  border-radius: 15px;
+
+  display: flex;
+  align-items: center;
+
+  margin-left: 50px;
+
+  img {
+    width: 14px;
+    height: 14px;
+    margin-left: 28px;
+  }
+
+  p {
+    margin-left: 26px;
+
+    font-family: "Cafe24Ssurround";
+    font-style: normal;
+    font-weight: 700;
+    font-size: 14px;
+    color: rgba(100, 100, 100, 0.5);
+  }
+
+  @media screen and (max-width: 1150px) {
+    margin-left: auto;
+  }
+
+  @media screen and (max-width: 760px) {
+    display: none;
   }
 `;
